requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
affine==2.4.0
    # via rasterio
aiohttp==3.9.5
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
antlr4-python3-runtime==4.9.3
    # via omegaconf
asciitree==0.3.3
    # via zarr
attrs==23.2.0
    # via
    #   aiohttp
    #   cads-api-client
    #   rasterio
cads-api-client==1.5.2
    # via cdsapi
cartopy==0.24.1
    # via deepwaters (pyproject.toml)
cdsapi==0.7.4
    # via deepwaters (pyproject.toml)
certifi==2024.7.4
    # via
    #   pyogrio
    #   pyproj
    #   rasterio
    #   requests
    #   sentry-sdk
cf-xarray==0.10.0
    # via deepwaters (pyproject.toml)
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   click-plugins
    #   cligj
    #   dask
    #   rasterio
    #   wandb
click-plugins==1.1.1
    # via rasterio
cligj==0.7.2
    # via rasterio
cloudpickle==3.0.0
    # via dask
colorama==0.4.6
    # via
    #   click
    #   tqdm
contourpy==1.2.1
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dask==2024.11.1
    # via
    #   deepwaters (pyproject.toml)
    #   xhistogram
    #   xskillscore
docker-pycreds==0.4.0
    # via wandb
fasteners==0.19
    # via zarr
filelock==3.13.1
    # via torch
fonttools==4.53.1
    # via matplotlib
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec==2024.2.0
    # via
    #   dask
    #   lightning
    #   pytorch-lightning
    #   torch
geopandas==1.0.1
    # via
    #   deepwaters (pyproject.toml)
    #   regionmask
gitdb==4.0.11
    # via gitpython
gitpython==3.1.43
    # via wandb
idna==3.7
    # via
    #   requests
    #   yarl
isimip-client==1.0.1
    # via deepwaters (pyproject.toml)
jinja2==3.1.3
    # via torch
joblib==1.4.2
    # via scikit-learn
kiwisolver==1.4.5
    # via matplotlib
lightning==2.4.0
    # via deepwaters (pyproject.toml)
lightning-utilities==0.11.3.post0
    # via
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
locket==1.0.0
    # via partd
markupsafe==2.1.5
    # via jinja2
matplotlib==3.9.2
    # via
    #   deepwaters (pyproject.toml)
    #   cartopy
    #   seaborn
mpmath==1.3.0
    # via sympy
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
multipledispatch==1.0.0
    # via pyjanitor
multiurl==0.3.2
    # via cads-api-client
natsort==8.4.0
    # via pyjanitor
networkx==3.2.1
    # via torch
numcodecs==0.12.1
    # via zarr
numpy==1.26.4
    # via
    #   cartopy
    #   contourpy
    #   dask
    #   geopandas
    #   matplotlib
    #   numcodecs
    #   pandas
    #   patsy
    #   properscoring
    #   pyogrio
    #   pytorch-lightning
    #   rasterio
    #   regionmask
    #   rioxarray
    #   scikit-learn
    #   scipy
    #   seaborn
    #   shapely
    #   snuggs
    #   statsmodels
    #   torchmetrics
    #   xarray
    #   xhistogram
    #   xskillscore
    #   zarr
omegaconf==2.3.0
    # via deepwaters (pyproject.toml)
packaging==24.1
    # via
    #   cartopy
    #   dask
    #   geopandas
    #   lightning
    #   lightning-utilities
    #   matplotlib
    #   pooch
    #   pyogrio
    #   pytorch-lightning
    #   regionmask
    #   rioxarray
    #   statsmodels
    #   torchmetrics
    #   xarray
pandas==2.2.2
    # via
    #   geopandas
    #   pandas-flavor
    #   seaborn
    #   statsmodels
    #   xarray
pandas-flavor==0.6.0
    # via pyjanitor
partd==1.4.2
    # via dask
patsy==0.5.6
    # via statsmodels
pillow==10.4.0
    # via matplotlib
platformdirs==4.2.2
    # via
    #   pooch
    #   wandb
pooch==1.8.2
    # via regionmask
properscoring==0.1
    # via xskillscore
protobuf==5.27.2
    # via wandb
psutil==6.0.0
    # via wandb
pyjanitor==0.29.2
    # via deepwaters (pyproject.toml)
pyogrio==0.9.0
    # via geopandas
pyparsing==3.1.2
    # via
    #   matplotlib
    #   snuggs
pyproj==3.6.1
    # via
    #   cartopy
    #   geopandas
    #   rioxarray
pyshp==2.3.1
    # via cartopy
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   multiurl
    #   pandas
pytorch-lightning==2.3.3
    # via lightning
pytz==2024.1
    # via
    #   multiurl
    #   pandas
pyyaml==6.0.1
    # via
    #   dask
    #   lightning
    #   omegaconf
    #   pytorch-lightning
    #   wandb
rasterio==1.3.9
    # via
    #   regionmask
    #   rioxarray
regionmask==0.12.1
    # via deepwaters (pyproject.toml)
requests==2.32.3
    # via
    #   cads-api-client
    #   cdsapi
    #   isimip-client
    #   multiurl
    #   pooch
    #   wandb
rioxarray==0.18.1
    # via deepwaters (pyproject.toml)
ruamel-yaml==0.18.6
    # via deepwaters (pyproject.toml)
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
scikit-learn==1.5.2
    # via deepwaters (pyproject.toml)
scipy==1.14.0
    # via
    #   properscoring
    #   pyjanitor
    #   scikit-learn
    #   statsmodels
    #   xskillscore
seaborn==0.13.2
    # via deepwaters (pyproject.toml)
sentry-sdk==2.8.0
    # via wandb
setproctitle==1.3.3
    # via wandb
setuptools==75.4.0
    # via
    #   lightning-utilities
    #   rasterio
    #   torch
    #   wandb
shapely==2.0.4
    # via
    #   cartopy
    #   geopandas
    #   regionmask
six==1.16.0
    # via
    #   docker-pycreds
    #   patsy
    #   python-dateutil
smmap==5.0.1
    # via gitdb
snuggs==1.4.7
    # via rasterio
statsmodels==0.14.2
    # via xskillscore
sympy==1.13.1
    # via torch
tabulate==0.9.0
    # via deepwaters (pyproject.toml)
threadpoolctl==3.5.0
    # via scikit-learn
toolz==0.12.1
    # via
    #   dask
    #   partd
torch==2.5.1
    # via
    #   deepwaters (pyproject.toml)
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
torchinfo==1.8.0
    # via deepwaters (pyproject.toml)
torchmetrics==1.4.0.post0
    # via
    #   lightning
    #   pytorch-lightning
tqdm==4.67.0
    # via
    #   deepwaters (pyproject.toml)
    #   cdsapi
    #   lightning
    #   multiurl
    #   pytorch-lightning
typing-extensions==4.12.2
    # via
    #   cads-api-client
    #   lightning
    #   lightning-utilities
    #   pytorch-lightning
    #   torch
tzdata==2024.1
    # via pandas
urllib3==2.2.2
    # via
    #   requests
    #   sentry-sdk
wandb==0.18.6
    # via deepwaters (pyproject.toml)
xarray==2024.10.0
    # via
    #   deepwaters (pyproject.toml)
    #   cf-xarray
    #   pandas-flavor
    #   regionmask
    #   rioxarray
    #   xhistogram
    #   xskillscore
xhistogram==0.3.2
    # via xskillscore
xskillscore==0.0.26
    # via deepwaters (pyproject.toml)
yarl==1.9.4
    # via aiohttp
zarr==2.18.3
    # via deepwaters (pyproject.toml)
